Fixed issue with throwing 500 Internal errors

+++ src/main/java/npc/martin/aomsbackend/apis/AnimalProgressRecordsRestController.java
@@ -8,6 +8,7 @@ import npc.martin.aomsbackend.services.AnimalProgressRecordsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -25,6 +26,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api")
+@CrossOrigin(origins = { "http://localhost:5173", "http://127.0.0.1:5173" } )
 public class AnimalProgressRecordsRestController {
     public AnimalProgressRecordsService animalProgressRecordsService;
 
@@ -37,62 +39,52 @@ public class AnimalProgressRecordsRestController {
     }
     
     @GetMapping(value = "/animals/getAnimalRecord/{animalId}")
-    public ResponseEntity<AnimalProgressRecords> getAnimalRecord(@PathVariable Integer animalId) throws Exception {
-        try {
-            Optional<AnimalProgressRecords> theRecordsContainer = animalProgressRecordsService.getAnimalDetails(animalId);
-            
-            if(theRecordsContainer.isPresent()) {
-                return new ResponseEntity<>(theRecordsContainer.get(), HttpStatus.FOUND);
-            } else {
-                throw new ResourceNotFoundException("Animal progress records for the id: " + animalId + "do not exist.");
-            }
-        } catch(Exception ex) {
-            throw new Exception("Encountered an error trying to serve the request.");
+    public ResponseEntity<AnimalProgressRecords> getAnimalRecord(@PathVariable Integer animalId) {
+        Optional<AnimalProgressRecords> theRecordsContainer = animalProgressRecordsService.getAnimalDetails(animalId);
+
+        if(theRecordsContainer.isPresent()) {
+            return new ResponseEntity<>(theRecordsContainer.get(), HttpStatus.FOUND);
+        } else {
+            throw new ResourceNotFoundException("Animal progress records for the id: " + animalId + "do not exist.");
         }
     }
     
     @PostMapping(value = "/animals/newAnimalRecord")
-    public ResponseEntity<AnimalProgressRecords> createAnimalRecord(@RequestBody AnimalProgressRecords animalProgressRecord) throws Exception {
-        try {
-            AnimalProgressRecords createdRecords = animalProgressRecordsService.createAnimalProgressRecord(animalProgressRecord);
-            return new ResponseEntity<>(createdRecords, HttpStatus.CREATED);
-        } catch(Exception ex) {
-            throw new Exception("Encountered an error trying to serve the request.");
-        }
+    public ResponseEntity<AnimalProgressRecords> createAnimalRecord(@RequestBody AnimalProgressRecords animalProgressRecord) {
+        AnimalProgressRecords createdRecords = animalProgressRecordsService.createAnimalProgressRecord(animalProgressRecord);
+        return new ResponseEntity<>(createdRecords, HttpStatus.CREATED);
     }
     
     @PutMapping("/animals/updateAnimalRecord")
-    public ResponseEntity<AnimalProgressRecords> updateAnimalRecord(@RequestBody AnimalProgressRecords animalProgressRecord) throws Exception {
-        try {
-            Optional<AnimalProgressRecords> theRecordsContainer = animalProgressRecordsService.getAnimalDetails(animalProgressRecord.getAnimalId());
-            AnimalProgressRecords currentAnimalProgressRecord = null;
-            
-            if(theRecordsContainer.isPresent()) {
-                currentAnimalProgressRecord = theRecordsContainer.get();
-            } else {
-                throw new ResourceNotFoundException("Animal progress records for the id: " + animalProgressRecord.getAnimalId() + "do not exist.");
-            }
-            
-            currentAnimalProgressRecord.setClinicalState(animalProgressRecord.getClinicalState());
-            currentAnimalProgressRecord.setFoods(animalProgressRecord.getFoods());
-            currentAnimalProgressRecord.setMedicalProgress(animalProgressRecord.getMedicalProgress());
-            currentAnimalProgressRecord.setWhereFound(animalProgressRecord.getWhereFound());
-            
-            animalProgressRecordsService.updateAnimalProgressRecord(currentAnimalProgressRecord);
-            return new ResponseEntity<>(currentAnimalProgressRecord, HttpStatus.OK);
-        } catch(Exception ex) {
-            throw new Exception("Encountered an error trying to serve the request.");
+    public ResponseEntity<AnimalProgressRecords> updateAnimalRecord(@RequestBody AnimalProgressRecords animalProgressRecord) {
+        Optional<AnimalProgressRecords> theRecordsContainer = animalProgressRecordsService.getAnimalDetails(animalProgressRecord.getAnimalId());
+        AnimalProgressRecords currentAnimalProgressRecord = null;
+
+        if(theRecordsContainer.isPresent()) {
+            currentAnimalProgressRecord = theRecordsContainer.get();
+        } else {
+            throw new ResourceNotFoundException("Animal progress records for the id: " + animalProgressRecord.getAnimalId() + "do not exist.");
         }
+
+        currentAnimalProgressRecord.setClinicalState(animalProgressRecord.getClinicalState());
+        currentAnimalProgressRecord.setFoods(animalProgressRecord.getFoods());
+        currentAnimalProgressRecord.setMedicalProgress(animalProgressRecord.getMedicalProgress());
+        currentAnimalProgressRecord.setWhereFound(animalProgressRecord.getWhereFound());
+
+        animalProgressRecordsService.updateAnimalProgressRecord(currentAnimalProgressRecord);
+        return new ResponseEntity<>(currentAnimalProgressRecord, HttpStatus.OK);
     }
     
     @DeleteMapping("/animals/deleteAnimalRecord/{animalId}")
-    public ResponseEntity<String> deleteAnimalRecord(@PathVariable Integer animalId) throws Exception {
-        try {
-            animalProgressRecordsService.deleteAnimalProgressRecord(animalId);
-            return new ResponseEntity<>("Successfully deleted records for animal with id: " + animalId, HttpStatus.OK);
-        } catch(Exception ex) {
-            throw new Exception("Encountered an error trying to serve the request.");
+    public ResponseEntity<String> deleteAnimalRecord(@PathVariable Integer animalId) {
+        Optional<AnimalProgressRecords> theRecordsContainer = animalProgressRecordsService.getAnimalDetails(animalId);
+        
+        if(theRecordsContainer.isEmpty()) {
+            throw new ResourceNotFoundException("No animal for id: " + animalId + " found.");
         }
+        
+        animalProgressRecordsService.deleteAnimalProgressRecord(animalId);
+        return new ResponseEntity<>("Successfully deleted records for animal with id: " + animalId, HttpStatus.OK);
     }
     
 }
